The final commit
-Delete IDispose in transaction ,
-Style Cop for unit test

+++ DrinkWater/DrinkWater/Model/Dailystatistic.cs
@@ -1,12 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 
+#nullable disable
+
 namespace DrinkWater
 {
     public partial class Dailystatistic
     {
         public int? UserIdRef { get; set; }
         public long? FluidIdRef { get; set; }
-        public decimal? Sum { get; set; }
+        public double? Sum { get; set; }
     }
 }
